Remove old photo only if it was from utfs and set it up correctly in the db

+++ app/api/uploadthing/core.ts
@@ -35,19 +35,20 @@ export const ourFileRouter = {
     })
     .onUploadComplete(async ({ metadata, file }) => {
       const oldImageUrl = metadata.profile.imageUrl;
+      const imageUrl = file.url;
       const res = await updateProfile(metadata.profile.id, {
-        imageUrl: metadata.profile.imageUrl,
+        imageUrl,
       });
 
       if (res.success) {
-        if (oldImageUrl) {
+        if (oldImageUrl && oldImageUrl.includes("utfs")) {
           const [_, key] = oldImageUrl.split("/f/");
           await new UTApi().deleteFiles(key);
         }
 
         return {
           ...res,
-          imageUrl: file.url,
+          imageUrl,
         };
       }
